product record handling

+++ resources/product/products.py
@@ -19,18 +19,23 @@ class ClsProduct(Resource):
     def get(self):
         try:
             conn = mysql.connector.connect(**config)
+            cursor = conn.cursor(dictionary=True)
             product_id = int(request.headers.get('product_id'))
-            if not isinstance(conn, str):
-                cursor = conn.cursor(dictionary=True)
-                query = "SELECT product_id, product_name, product_status FROM product \
-                         WHERE product_id = {product_id}".format(product_id = product_id)
-                cursor.execute(query)
-                productRow = cursor.fetchone()
-                close_connection(conn, cursor)
+            cursor.execute("SELECT count(*) FROM product WHERE product_id=%s", (product_id,))
+            check_record_exist = cursor.fetchone().get('count(*)')
+            if check_record_exist:
+                if not isinstance(conn, str):
+                    query = "SELECT product_id, product_name, product_status FROM product \
+                             WHERE product_id = {product_id}".format(product_id = product_id)
+                    cursor.execute(query)
+                    productRow = cursor.fetchone()
+                    close_connection(conn, cursor)
 
-                return api_success(productRow, "Products details fetched successfully")
+                    return api_success(productRow, "Products details fetched successfully")
+                else:
+                    return api_failure(str(conn))
             else:
-                return api_failure(str(conn))
+                return { "message": "Record not found" }, 404
         except Exception as error:
             return api_failure(str(error))
 
@@ -60,22 +65,27 @@ class ClsProduct(Resource):
     def put(self):
         try:
             conn = mysql.connector.connect(**config)
+            cursor = conn.cursor(dictionary=True)
             _json = request.json
             _product_id = _json['product_id']
             _product_name = _json['product_name']
             _product_status = _json['product_status']
-            if not isinstance(conn, str):
-                cursor = conn.cursor(dictionary=True)
-                query = "UPDATE product set product_name=%s, product_status=%s where product_id = %s"
-                bindData = (_product_name, _product_status, _product_id)
-                cursor = conn.cursor()
-                cursor.execute(query, bindData)
-                conn.commit()
-                close_connection(conn, cursor)
+            cursor.execute("SELECT count(*) FROM product WHERE product_id=%s", (_product_id,))
+            check_record_exist = cursor.fetchone().get('count(*)')
+            if check_record_exist:
+                if not isinstance(conn, str):
+                    query = "UPDATE product set product_name=%s, product_status=%s where product_id = %s"
+                    bindData = (_product_name, _product_status, _product_id)
+                    cursor = conn.cursor()
+                    cursor.execute(query, bindData)
+                    conn.commit()
+                    close_connection(conn, cursor)
 
-                return api_success(None, "Product updated successfully")
+                    return api_success(None, "Product updated successfully")
+                else:
+                    return api_failure(str(conn))
             else:
-                return api_failure(str(conn))
+                return { "message": "Record not found" }, 404
         except Exception as error:
             return api_failure(str(error))
 
@@ -83,17 +93,22 @@ class ClsProduct(Resource):
     def delete(self):
         try:
             conn = mysql.connector.connect(**config)
+            cursor = conn.cursor(dictionary=True)
             _product_id = int(request.headers.get('product_id'))
-            if not isinstance(conn, str):
-                cursor = conn.cursor(dictionary=True)
-                query = "DELETE FROM product WHERE product_id =%s"
-                bindData = (_product_id, )
-                cursor.execute(query, bindData)
-                conn.commit()
-                close_connection(conn, cursor)
+            cursor.execute("SELECT count(*) FROM product WHERE product_id=%s", (_product_id,))
+            check_record_exist = cursor.fetchone().get('count(*)')
+            if check_record_exist:
+                if not isinstance(conn, str):
+                    query = "DELETE FROM product WHERE product_id =%s"
+                    bindData = (_product_id, )
+                    cursor.execute(query, bindData)
+                    conn.commit()
+                    close_connection(conn, cursor)
 
-                return api_success(None, "Products deleted successfully")
+                    return api_success(None, "Products deleted successfully")
+                else:
+                    return api_failure(str(conn))
             else:
-                return api_failure(str(conn))
+                return { "message": "Record not found" }, 404
         except Exception as error:
             return api_failure(str(error))
